Split into batching.py, fix tests
