+++ nautilus_trader/backtest/node.py
@@ -39,6 +39,8 @@ from nautilus_trader.model.identifiers import Venue
 from nautilus_trader.model.instruments.base import Instrument
 from nautilus_trader.model.objects import Money
 from nautilus_trader.model.orderbook.data import OrderBookDelta
+from nautilus_trader.persistence.batching import generate_data_batches
+from nautilus_trader.persistence.catalog import DataCatalog
 from nautilus_trader.trading.config import ImportableStrategyConfig
 from nautilus_trader.trading.config import StrategyFactory
 from nautilus_trader.trading.strategy import TradingStrategy
@@ -279,10 +281,10 @@ def streaming_backtest_runner(
     batch_size_bytes: Optional[int] = None,
 ):
     config = data_configs[0]
-    catalog = config.catalog()
-
-    streaming_kw = merge_data_configs_for_calc_streaming_chunks(data_configs=data_configs)
-    for start, end in catalog.calc_streaming_chunks(**streaming_kw, target_size=batch_size_bytes):
+    catalog: DataCatalog = config.catalog()
+    for start, end in generate_data_batches(
+        catalog=catalog, data_configs=data_configs, batch_size=batch_size_bytes
+    ):
         engine.clear_data()
         for config in data_configs:
             data = config.load(start_time=start, end_time=end)
@@ -296,23 +298,6 @@ def streaming_backtest_runner(
 # Register tokenization methods with dask
 for cls in Instrument.__subclasses__():
     normalize_token.register(cls, func=cls.to_dict)
-
-
-def merge_data_configs_for_calc_streaming_chunks(data_configs: List[BacktestDataConfig]):
-    instrument_ids = [c.instrument_id for c in data_configs]
-    data_types = [c.data_type for c in data_configs]
-    starts = [c.start_time for c in data_configs]
-    if len(set(starts)) > 1:
-        logger.warning("Multiple start dates in data_configs, using min")
-    ends = [c.end_time for c in data_configs]
-    if len(set(ends)) > 1:
-        logger.warning("Multiple start dates in data_configs, using min")
-    return {
-        "instrument_ids": instrument_ids,
-        "data_types": data_types,
-        "start_time": starts[0],
-        "end_time": ends[0],
-    }
 
 
 @normalize_token.register(object)
